Deleted RequestFactory, added createElevatorRequest to InputFileReader
